requirements.txt
asgiref==3.8.1
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
pip==23.2.1
pymongo==3.12.1
pytz==2024.2
setuptools==68.2.0
sqlparse==0.2.4
tzdata==2024.2
wheel==0.41.2
djangorestframework==3.15.2